refactor(state): split state_test.go into focused test files (#188)
